Initial Upload
Here's everything I've created so far.
oSDL is an objectified wrapper for SDL so that it flows better in the
c++ syntax. It also includes many convenience features I feel will
expedite game and app development.

I have the rendering features encapsulated as well as the basics of
window handling. Next I'm writing some convenience classes that will
auto-manage windows for the user so that they do not need to worry about
the event loop at all.

The next step after is audio.